Fixed header inclusion on case sensitive systems
Linux, by default, uses a case sensitive filesystem.  OS X can be enabled to use a case sensitive filesystem.

If you include a file you MUST ensure that the case in the #include EXACTLY matches the case of the filesystem otherwise it will be BROKEN on anything but Windows and a default unmodified OS X installation.